9C-253 Refactoring Interpreters
We perform some refactorings to the interpreters of web tasks.
First, we introduce a set of new type alias, `QueryRequest`,
`QueryResult`, and `QueryService`.
We transform `ApiClient` and `WebScrapper` to functions (services).

We refactor the services package. We transform the ApiClient and
WebScrapper into functions, since it is their use as external functions
what is more relevant from there.

Also, we activate a compiler warning in the build: unused imports