9C-253 Implement Cached Query Service
We add two classes: `RedisCachedMonadicFunction` and
`CachedQueryService`.

`RedisCachedMonadicFunction` represents the caching of functions that
return monadic values. This class is parametric in the Monadic type
`M[_]` in which it gives the results. The class is abstract, since
it defines abstract type members for the type of Keys and Values stored
in the Redis cache.

`CachedQueryService` specialises (inherits) `RedisCachedMonadicFunction`
for the Query Service. This class takes as a constructor parameter the
 name of the service, which is used to mark the keys in the cache.

The keys in the cache only store the name of the package and the
localization, since it is only these that are used to read a package.