9C-626 Refactor collections lists endpoints
The API has several endpoints that return a list of shared collections.
Until now, the response body from those endpoints (in success) is a Json
Array, where each element is the Json object of a shared collection.
We change the response, so the Json Array is wrapped in a Json object,
under a field called "collections".

The affected endpoints are the following ones:
* GET /collections
* GET /collections/latest/{category}/{pageNumber}/{pageSize}
* GET /collections/top/{category}/{pageNumber}/{pageSize}