9C-613 Refactor Routes, Marshallers.
We introduce a Marshaller Factory, that for a (High-Kind) type F[_],
an object that makes as new `ToResponseMarshaller[F[A]]` for each A.
This saves us from having to specify a marshaller for each object.
We add specific instances for Task and Id.

We add a Marshaller factory that uses a `Monad` instance and a natural
transformation to move across Hight-King types `F[A]` and `G[A]`.

We refactor NineCardsRoutes. Now it is a class instead of an object and
def, and now we can split the sub routes into separate variables.