9C-692 Introduce FullCard domain class.
We introduce a new domain class, FullCard, which stores all the
information that we need to create an app's card, or recommendation,
and which should be stored in the cache.

* In the API, we add the messages `ApiCard`, `ApiCardList`,
  `ApiRecommendation`, `ApiRecommendationList` that answer each
  endpoint.
* In the `domain`, we add the class `FullCard`.
* In the `services`, we replace any use of `AppCard` or
  `AppRecommendation` by `FullCard`.

In the Web Scraper Converters, we generalise the method to get an Icon,
so we can also use it to get screenshots.