695 Implement workflow to refresh cache.
We implement a new workflow process, that retrieves from the
cache a (given) number of pending packages, and tries to solve them.

In the Web Scrapper Service, we add a new operation to get a card.
In the Cache Service, we add new operations to unmarl a package as
pending, and to get the list of cache keys with pending status.
We create two top-level processes with cards, one for resolving a
package, and another one to searching and resolving `N` packages.