9C-834 Integrate new workflow into application
We integrate the new workflow into the processes for the endpoints of
getting cards, getting recommendations by category, and getting
recommendations by a list of applications.

We modify the ApiServices: we add it the new AppServiceByProcess
parameter, and connect all the methods through there.
For testing, we transform Wiring into a class, so that each test may
perform a different shutdown process.