Web Scraper: repair error noted in QA.
During QA of ticket on redis cache improvements, it was noted
that for some apps (such as Google Sky Map), the web page contains
as screenshot or icon an url with the `https` beginning, and fail.
During QA, an error was noticed involving a match error when using
matching against regular expressions. We fix it.

We also add a unit test with the Sky Map web page, for testing the
fix.