9C-1039 OAuth Services
We create a new service that relates to Google's OAuth Servers.
We define and implement one operation in this service, to obtain a fresh
access token using the credentials of a Service Account.
We write some tests for the converter toGoogleCredential, for which
we have developed some special generators.
We have also moved some shared generators to the `domain` package.

Unfortunately, generating a private key (pkcs8-base64) is a bit too
complex, so for now we have hard-coded one of example.