9C-972 Add Endpoint to return title and Icon
We modify an existing endpoint from the API, the one about details, so
that it can be used to return only the title and icon of each app in a
list. We modify the Api Models, so that ApiAppsInfoResponse is
parametrized by the type of the info given about each App.