9C-1075 Cache Keys: avoid Json Encoding
In the original implementation, we used Json encoding of keys as a
simple and generic implementation. Unfortunately, this encoding
makes it difficult to inspect the contents of the cache.
Here, we use String formatting and Regex-based parsing for keys.

We use the semicolon for separate the parts of the key, on the
understanding that this is not a valid character.