9C-1077 Implement Resolve Apps Actor
We trim and rename the methods, in services and processes, for resolving
a number of apps from the pending package queue in the Redis cache.

We create a new scala package, `cards.nine.app`, in the `api` module,
in this package we put all things related to the running of the
application: the actors, loggers, and the main logger.