9C-1097 Replacing scala-redis with Scredis
We move the implementation of the Redis Cache commons pcakge, from
[Scala-Redis](https://github.com/debasishg/scala-redis), to use
the [scredis](https://github.com/scredis/scredis) client.
Scredis is an asynchronous non-blocking client.

Since `scredis` methods use `scala.concurrent.Future`, we add a
conversion between `scala.concurrent.Future`
to`scalaz.concurrent.Task`, from "http://stackoverflow.com/a/17377768"

We modify `CacheWrapper` and `CacheQueue`. Instead of its methods
carrying out side-effectful computations, now the methods return
_functions_ from a RedisClient (an object with some operations)
to a `Task`. We call this type a `RedisOps`. The _interpretation_
of a RedisOps is to execute the method as a Redis Transaction,
which is to say, a batch group of operations sent in a single round-trip.

We introduce an `Applicative` instance for `RedisOps`: the goal is to
join independent Redis operations or queries in a sinble batch work,
that uses a single round-trip. Note that a `Monad` instance would need
two round-trips for each flatMap.