9C-1212 GetDetailsList Operation
We create an operation GetDetailsList in the algebra of Google Play
Api operations. The semantics of this operation is to take a list of
package names, and perform the GetDetails operation in each of them.

Unlike the other operations in the algebra, the interpretation of
GetDetailsList can involve several  HTTP operations.
The interpretation splits the lists in several batches (sublists of
operations) and resolves those batches in sequence; each batch is
resolved in _parallel_, to reduce response time.

Between each batches, we check if any HTTP connection gave back a
401 Unauthorized or a 429 Too Many Requests error, in which case
we stop processing the rest of packages.

The part about has to be done in `services` instead of `processes`,
because the Processes only knows about the `Free` monadic type, so
we can not use the parallelism operations of `Task` in there.