9C-1212 Google Play Services.
As part of the ticket, we add a config parameter, for the maximum
parallelism in the HTTP client for the Api and the Web Scrapper.

To pass this parameter, we have to turn `Wiring` into a class.
Along the way, we split it, so the operations GooglePlay are
interpreted in two steps: first to inner algebra, and then to Task.